requirements.txt
numpy
scipy
scikit-learn
matplotlib
pyyaml
gym
mujoco-py
tensorflow
torch